refactor(obd2): extract the Mode 22 odometer parsers + shared decode helpers from elm327_parsers (#3279)
Second decomposition slice (after the #3278 VIN extraction). Moves the Mode 22
manufacturer-odometer parsers (parseMfgOdometer3Byte/2Byte/MilesTimes10,
_parseMode22Body, cleanResponse22) into elm327_mode22_parsers.dart, and the
shared low-level helpers (_parseHexBytes → parseElmHexBytes, the odometer
plausibility gate → isPlausibleOdometerKm) into elm327_decode_util.dart so a
single definition serves both parser files.

Behavior-preserving: Elm327Protocol's facade delegates repoint to the new
class (production callers use the facade and are untouched); only the direct
`Elm327Parsers.parseMfgOdometer*` / `cleanResponse22` test references move. The
grandfathered elm327_parsers.dart ratchets DOWN 519 → 419; the remaining Mode 01
bulk keeps it just over the 400 norm, so #3279 stays referenced for the last
slice.
